Move `repr/robdd.rs` -> `repr/bdd.rs`
Since the Ptr type itself doesn't actually know about the ordering!
