+++ examples/bayesian_network_compiler.rs
@@ -8,9 +8,9 @@ use rsdd::builder::bdd::builder::BddBuilder;
 use rsdd::builder::cache::all_app::AllTable;
 use rsdd::builder::decision_nnf_builder::DecisionNNFBuilder;
 use rsdd::builder::sdd::{builder::SddBuilder, compression::CompressionSddBuilder};
+use rsdd::repr::bdd::BddPtr;
 use rsdd::repr::ddnnf::DDNNFPtr;
 use rsdd::repr::dtree::DTree;
-use rsdd::repr::robdd::BddPtr;
 use rsdd::repr::var_order::VarOrder;
 use rsdd::repr::vtree::VTree;
 use rsdd::util::semirings::realsemiring::RealSemiring;
