+++ src/builder/sdd/compression.rs
@@ -7,8 +7,8 @@ use crate::builder::cache::all_app::AllTable;
 use crate::builder::cache::ite::Ite;
 use crate::builder::cache::LruTable;
 use crate::builder::BottomUpBuilder;
+use crate::repr::bdd::create_semantic_hash_map;
 use crate::repr::ddnnf::DDNNFPtr;
-use crate::repr::robdd::create_semantic_hash_map;
 use crate::repr::sdd::binary_sdd::BinarySDD;
 use crate::repr::sdd::sdd_or::{SddAnd, SddOr};
 use crate::repr::sdd::SddPtr;
